Refactor SearchRequest and SearchResponse classes
Make it so we can move to a world where all we have to do is call `Catalog.create(DataProvider.search(query="foo"))` to generated a paginated feed. `create` could take an argument `paginate=True` that uses the output of DataProvider.search, i.e. DataProvider.SearchResult.request

`Catalog.create` paginate requires the base_url from the provider, and so we should also add the `provider` to the `DataProvider.SearchResult`.